structure database mod improvement
