+++ database/src/base/state_indexer.rs
@@ -1,0 +1,109 @@
+#[async_trait::async_trait]
+pub trait StateIndexerDbManager {
+    async fn add_state_changes(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+        key: &[u8],
+        value: &[u8],
+    ) -> anyhow::Result<()>;
+
+    async fn delete_state_changes(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+        key: &[u8],
+    ) -> anyhow::Result<()>;
+
+    async fn add_access_key(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+        public_key: &[u8],
+        access_key: &[u8],
+    ) -> anyhow::Result<()>;
+
+    async fn delete_access_key(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+        public_key: &[u8],
+    ) -> anyhow::Result<()>;
+
+    #[cfg(feature = "account_access_keys")]
+    async fn get_access_keys(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+    ) -> anyhow::Result<std::collections::HashMap<String, Vec<u8>>>;
+
+    #[cfg(feature = "account_access_keys")]
+    async fn add_account_access_keys(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        public_key: &[u8],
+        access_key: Option<&[u8]>,
+    ) -> anyhow::Result<()>;
+
+    #[cfg(feature = "account_access_keys")]
+    async fn update_account_access_keys(
+        &self,
+        account_id: String,
+        block_height: u64,
+        account_keys: std::collections::HashMap<String, Vec<u8>>,
+    ) -> anyhow::Result<()>;
+
+    async fn add_contract_code(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+        code: &[u8],
+    ) -> anyhow::Result<()>;
+
+    async fn delete_contract_code(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+    ) -> anyhow::Result<()>;
+
+    async fn add_account(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+        account: Vec<u8>,
+    ) -> anyhow::Result<()>;
+
+    async fn delete_account(
+        &self,
+        account_id: near_indexer_primitives::types::AccountId,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+    ) -> anyhow::Result<()>;
+
+    async fn add_block(
+        &self,
+        block_height: u64,
+        block_hash: near_indexer_primitives::CryptoHash,
+    ) -> anyhow::Result<()>;
+
+    async fn add_chunks(
+        &self,
+        block_height: u64,
+        chunks: Vec<(
+            crate::primitives::ChunkHash,
+            crate::primitives::ShardId,
+            crate::primitives::HeightIncluded,
+        )>,
+    ) -> anyhow::Result<()>;
+
+    async fn update_meta(&self, indexer_id: &str, block_height: u64) -> anyhow::Result<()>;
+    async fn get_last_processed_block_height(&self, indexer_id: &str) -> anyhow::Result<u64>;
+}
