+++ database/src/lib.rs
@@ -4,44 +4,16 @@ use crate::base::BaseDbManager;
 pub use crate::base::ReaderDbManager;
 pub use crate::base::StateIndexerDbManager;
 pub use crate::base::TxIndexerDbManager;
+
 #[cfg(feature = "scylla_db")]
 mod scylladb;
 #[cfg(feature = "scylla_db")]
-pub(crate) use crate::scylladb::base::ScyllaStorageManager;
+pub use crate::scylladb::AdditionalDatabaseOptions;
+
+#[cfg(not(feature = "scylla_db"))]
+pub use crate::base::AdditionalDatabaseOptions;
 
 pub mod primitives;
-
-/// Additional database options for scylla database
-///
-/// `preferred_dc: Option<String>`,
-/// ScyllaDB preferred DataCenter
-/// Accepts the DC name of the ScyllaDB to filter the connection to that DC only (preferrably).
-/// If you connect to multi-DC cluter, you might experience big latencies while working with the DB.
-/// This is due to the fact that ScyllaDB driver tries to connect to any of the nodes in the cluster
-/// disregarding of the location of the DC. This option allows to filter the connection to the DC you need.
-/// Example: "DC1" where DC1 is located in the same region as the application.
-///
-/// `keepalive_interval: Option<u64>`
-/// ScyllaDB keepalive interval
-///
-/// `max_retry: u8`
-/// Max retry count for ScyllaDB if `strict_mode` is `false`
-///
-/// `strict_mode: bool`
-/// Attempts to store data in the database should be infinite to ensure no data is missing.
-/// Disable it to perform a limited write attempts (`max_retry`)
-/// before skipping giving up and moving to the next piece of data
-#[cfg(feature = "scylla_db")]
-pub struct AdditionalDatabaseOptions {
-    pub preferred_dc: Option<String>,
-    pub keepalive_interval: Option<u64>,
-    pub max_retry: u8,
-    pub strict_mode: bool,
-}
-
-/// Additional database options for postgres database(WIP)
-#[cfg(not(feature = "scylla_db"))]
-pub struct AdditionalDatabaseOptions {}
 
 async fn prepare_db_manager<T>(
     database_url: &str,
@@ -77,7 +49,7 @@ pub async fn prepare_read_rpc_db_manager(
     .await
 }
 
-pub async fn prepare_stata_indexer_db_manager(
+pub async fn prepare_state_indexer_db_manager(
     database_url: &str,
     database_user: Option<&str>,
     database_password: Option<&str>,
